Move files to widget package
