+++ app/src/main/res/xml/dashboard_widget_info.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <appwidget-provider xmlns:android="http://schemas.android.com/apk/res/android"
-    android:configure="com.aspirebudgetingmobile.aspirebudgeting.WidgetConfigureActivity"
+    android:configure="com.aspirebudgetingmobile.aspirebudgeting.widget.WidgetConfigureActivity"
     android:initialKeyguardLayout="@layout/widget"
     android:initialLayout="@layout/widget"
     android:minWidth="180dp"
